refactor(mcp): drain workforce (HR) tool domain into a scoped pack (BI-2B7EE073)
Extract the six workforce/HR lifecycle tools — query_employees,
list_departments, list_positions, create_employee,
transition_employee_status, propose_leave_policy — out of the mcp-tools.ts
executeTool switch into apps/web/lib/mcp/packs/workforce-pack.ts. Definitions,
coercion, Prisma calls, envelopes, requiredCapability, and grants are moved
verbatim; each handler lazy-imports the Prisma client so the pack has zero
dependency on mcp-tools.ts internals. mcp-tools.ts registration appends the pack
to composeToolPacks. Ratchet baseline shrinks 212 -> 206 inline cases; module-size
baseline retightened (mcp-tools.ts 15446 -> 15078 LOC). New workforce-pack.test.ts
mirrors the pack precedents (tool names, provenance-free descriptions, grants
mirror, per-handler envelopes with a mocked Prisma client).

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the workforce tool domain into a scoped pack; ratchet baseline shrunk; behavior-preserving relocation.
