+++ scripts/mcp-tool-pack-baseline.json
@@ -33,7 +33,6 @@
   "create_digital_product",
   "create_ea_element",
   "create_ea_relationship",
-  "create_employee",
   "create_epic",
   "create_knowledge_article",
   "create_licensing_readiness_issue",
@@ -95,12 +94,10 @@
   "list_backlog_items",
   "list_build_activity_since",
   "list_customer_accounts",
-  "list_departments",
   "list_epics",
   "list_my_capability_needs",
   "list_opportunities",
   "list_patch_posture",
-  "list_positions",
   "list_project_directory",
   "list_quotes",
   "list_sandbox_files",
@@ -117,13 +114,11 @@
   "propose_decomposition",
   "propose_file_change",
   "propose_improvement",
-  "propose_leave_policy",
   "propose_skill_improvement",
   "provision_build_engine",
   "publish_to_linkedin",
   "publish_wiki_overlay_pages",
   "query_backlog",
-  "query_employees",
   "query_ontology_graph",
   "query_version_history",
   "read_codebase_manifest",
@@ -194,7 +189,6 @@
   "summon_coworker",
   "tick_marketing_scheduler",
   "trace_code_surface",
-  "transition_employee_status",
   "triage_backlog_item",
   "trigger_contributor_inventory_sync",
   "update_backlog_item",
